Override nettest and submariner-metrics-proxy
To ensure that we're testing with the latest image, as available in
the local repository, we need to override all users of the nettest
image. Currently subctl exposes two components for that:
submariner-nettest and submariner-metrics-proxy. This patch adds
support for multiple components per image; since the project already
preloads the nettest image, this results in the correct image override
arguments being given to subctl.

Signed-off-by: Stephen Kitt <skitt@redhat.com>